+++ src/content/services/index.ts
@@ -1,0 +1,3 @@
+export { observeDOM } from './DOMObserver'
+export { isRedminePage, findTextareas, hideToolbars, isTextareaInContext } from './RedmineService'
+export { canProcess, processTextarea, cleanupTextarea, cleanupAll } from './TextareaProcessor'
