+++ src/content/utils/index.ts
@@ -1,0 +1,8 @@
+export { DOMUtils } from './dom'
+export { 
+  REDMINE_SELECTORS, 
+  OVERLAY_CONFIG, 
+  MARKDOWN_OVERLAY_ATTRIBUTE, 
+  SYNC_INTERVAL_MS, 
+  CURSOR_SYNC_DELAY_MS 
+} from './constants'
